+++ R/check_scale.R
@@ -7,22 +7,19 @@
 #'
 #' @return integer scale of map
 check_scale <- function(x) {
-  if (is.numeric(x) && length(x) == 1) {
-    if (x %in% c(110, 50, 10)) {
+  if (is.numeric(x) && length(x) == 1L) {
+    if (x %in% c(110L, 50L, 10L)) {
       return(x)
     }
-  } else if (is.character(x) && length(x) == 1) {
-    xnew <- c(small = 110, medium = 50, large = 10)[tolower(x)]
+  } else if (is.character(x) && length(x) == 1L) {
+    xnew <- c(small = 110L, medium = 50L, large = 10L)[tolower(x)]
 
     if (!is.na(xnew)) {
       return(unname(xnew))
     }
   }
 
-  stop(
-    "Invalid `scale`. Must be one of [110, 50, 10, 'small', 'medium', 'large'] you have :",
-    x,
-    "\n",
-    call. = FALSE
+  cli::cli_abort(
+    "Invalid {.arg scale} argument. Must be one of [110, 50, 10, 'small', 'medium', 'large'] you provided : {.val {x}}."
   )
 }